# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Projecte Integrador API Documentation",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/patients/{patientId}/contacts": {
            "get": {
                "tags": [
                    "Contactos"
                ],
                "summary": "Lista todos los contactos de un paciente",
                "description": "Obtiene una lista de todos los contactos asociados a un paciente específico.",
                "operationId": "6c9d77d673adc16fed1594f7fb2f9d63",
                "parameters": [
                    {
                        "name": "patientId",
                        "in": "path",
                        "description": "ID del paciente",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operación exitosa",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ContactPerson"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                }
            },
            "post": {
                "tags": [
                    "Contactos"
                ],
                "summary": "Crea un nuevo contacto para un paciente",
                "description": "Crea un nuevo contacto asociado a un paciente específico.",
                "operationId": "d18c5cdf22e0bd701f67d776eb120532",
                "parameters": [
                    {
                        "name": "patientId",
                        "in": "path",
                        "description": "ID del paciente",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreContactRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Operación exitosa",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ContactPerson"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Solicitud incorrecta"
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                }
            }
        },
        "/api/contacts/{id}": {
            "put": {
                "tags": [
                    "Contactos"
                ],
                "summary": "Actualiza un contacto por ID",
                "description": "Actualiza la información de un contacto específico por su ID.",
                "operationId": "30863a9d3758a8c86d3c6d1fb000f800",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del contacto",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateContactRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Operación exitosa",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ContactPerson"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Solicitud incorrecta"
                    },
                    "404": {
                        "description": "Contacto no encontrado"
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Contactos"
                ],
                "summary": "Elimina un contacto por ID",
                "description": "Elimina un contacto específico por su ID.",
                "operationId": "4522941f7946c9d48b928cf6b9c5aa8a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del contacto",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Operación exitosa (sin contenido)"
                    },
                    "404": {
                        "description": "Contacto no encontrado"
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "StoreContactRequest": {
                "title": "Store Contact Request",
                "description": "Store Contact request body data",
                "required": [
                    "nombre",
                    "apellido",
                    "telefono",
                    "relacion"
                ],
                "properties": {
                    "nombre": {
                        "description": "Contact name",
                        "type": "string",
                        "maxLength": 255
                    },
                    "apellido": {
                        "description": "Contact last name",
                        "type": "string",
                        "maxLength": 255
                    },
                    "telefono": {
                        "description": "Contact phone number",
                        "type": "string",
                        "maxLength": 20
                    },
                    "relacion": {
                        "description": "Relationship to patient",
                        "type": "string",
                        "maxLength": 255
                    }
                },
                "type": "object"
            },
...
            "UpdateContactRequest": {
                "title": "Update Contact Request",
                "description": "Update Contact request body data",
                "properties": {
                    "nombre": {
                        "description": "Contact name",
                        "type": "string",
                        "maxLength": 255
                    },
                    "apellido": {
                        "description": "Contact last name",
                        "type": "string",
                        "maxLength": 255
                    },
                    "telefono": {
                        "description": "Contact phone number",
                        "type": "string",
                        "maxLength": 20
                    },
                    "relacion": {
                        "description": "Relationship to patient",
                        "type": "string",
                        "maxLength": 255
                    }
                },
                "type": "object"
            },
...
            "ContactPerson": {
                "title": "ContactPerson",
                "description": "ContactPerson model",
                "properties": {
                    "id": {
                        "description": "ContactPerson ID",
                        "type": "integer",
                        "format": "int64"
                    },
                    "nombre": {
                        "description": "ContactPerson name",
                        "type": "string"
                    },
                    "apellido": {
                        "description": "ContactPerson last name",
                        "type": "string"
                    },
                    "telefono": {
                        "description": "ContactPerson phone number",
                        "type": "string"
                    },
                    "relacion": {
                        "description": "Relationship to patient",
                        "type": "string"
                    },
                    "paciente_id": {
                        "description": "Patient ID",
                        "type": "integer",
                        "format": "int64"
                    },
                    "created_at": {
                        "description": "Creation timestamp",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Update timestamp",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}